+++ spec/fixtures/requests/stop_snoring.json
@@ -1,0 +1,51 @@
+{
+  "version": "1.0",
+  "session": {
+    "new": true,
+    "sessionId": "amzn1.echo-api.session.ID",
+    "application": {
+      "applicationId": "amzn1.ask.skill.ID"
+    },
+    "user": {
+      "userId": "amzn1.ask.account.ID"
+    }
+  },
+  "context": {
+    "AudioPlayer": {
+      "playerActivity": "IDLE"
+    },
+    "Display": {
+    },
+    "System": {
+      "application": {
+        "applicationId": "amzn1.ask.skill.ID"
+      },
+      "user": {
+        "userId": "amzn1.ask.account.USERID"
+      },
+      "device": {
+        "deviceId": "amzn1.ask.device.ID",
+        "supportedInterfaces": {
+          "AudioPlayer": {
+          },
+          "Display": {
+            "templateVersion": "1.0",
+            "markupVersion": "1.0"
+          }
+        }
+      },
+      "apiEndpoint": "https://api.amazonalexa.com",
+      "apiAccessToken": "API_ACCESS_TOKEN"
+    }
+  },
+  "request": {
+    "type": "IntentRequest",
+    "requestId": "amzn1.echo-api.request.ID",
+    "timestamp": "2018-04-02T15:04:42Z",
+    "locale": "en-US",
+    "intent": {
+      "name": "StopSnoring",
+      "confirmationStatus": "NONE"
+    }
+  }
+}
