+++ spec/fixtures/response/emergency.json
