+++ spec/fixtures/response/error.json
@@ -1,0 +1,12 @@
+{
+  "body": {
+    "version": "1.0",
+    "response": {
+      "outputSpeech": {
+        "type": "PlainText",
+        "text": "I did not understand you"
+      },
+      "shouldEndSession": true
+    }
+  }
+}
