+++ spec/fixtures/response/next_ten_minutes.json
