+++ spec/fixtures/response/stop_snoring.json
