+++ spec/fixtures/response/unauthorized.json
@@ -1,0 +1,12 @@
+{
+  "body": {
+    "version": "1.0",
+    "response": {
+      "outputSpeech": {
+        "type": "PlainText",
+        "text": "You can't make me! You're not my real dad!"
+      },
+      "shouldEndSession": true
+    }
+  }
+}
