+++ spec/spec_helper.rb
@@ -11,6 +11,8 @@ $LOAD_PATH.unshift(File.dirname(__FILE__))
   Dir[File.join(File.dirname(__FILE__), 'spec', dir, '/**/*.rb')].each { |f| require f }
 end
 
+WebMock.disable_net_connect!
+
 RSpec.configure do |config|
   config.expect_with :rspec do |expectations|
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
